python-services/ml-matching/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-core==2.27.1
supabase==2.10.0
postgrest==0.18.0
numpy==2.2.0
python-dotenv==1.0.1
httpx==0.27.2
